fix: detail panel cut off on desktop; numeric M/D dates
- ContactDetailPanel: md:relative re-activated the base top-[4.75rem] offset,
  pushing the panel below the viewport so its bottom/scroll were unreachable.
  Add md:top-0 to neutralize it on desktop.
- Follow-up / overdue / upcoming dates now render numeric M/D (e.g. 6/16)
  instead of 'Jun 16'.

+++ src/lib/upcoming.ts
@@ -9,7 +9,8 @@ function meetingShort(iso: string): string {
   return new Date(iso).toLocaleString(undefined, { month: 'short', day: 'numeric', hour: 'numeric', minute: '2-digit' });
 }
 function dateShort(iso: string): string {
-  return new Date(iso).toLocaleDateString(undefined, { month: 'short', day: 'numeric' });
+  // Numeric M/D (e.g. "6/16") for follow-up / overdue / upcoming dates.
+  return new Date(iso).toLocaleDateString(undefined, { month: 'numeric', day: 'numeric' });
 }
 
 /**
